requirements.txt
Django==4.2.7
psycopg2-binary==2.9.7
python-decouple==3.8
Pillow==10.0.1
django-crispy-forms==2.1
crispy-bootstrap4==2022.1
django-filter==23.3
reportlab==4.0.4
openpyxl==3.1.2
django-extensions==3.2.3
whitenoise==6.6.0
gunicorn==21.2.0
